refactor (schedulers/event): breakdown large file of scheduler into smaller ones
