+++ internal/schedulers/event/scheduler/worker/dispatch.go
@@ -1,0 +1,30 @@
+package worker
+
+import (
+	"time"
+
+	"github.com/ethereum/go-ethereum/core/types"
+)
+
+// performActionExecution handles the actual action execution logic
+func (w *EventWorker) performActionExecution(log types.Log) bool {
+	// TODO: Implement actual action execution logic
+	// This should:
+	// 1. Parse event data if needed
+	// 2. Send task to manager/keeper for execution
+	// 3. Handle response and update job status
+
+	// Simulate action execution for now
+	w.Logger.Info("Simulating action execution",
+		"job_id", w.Job.JobID,
+		"target_chain", w.Job.TargetChainID,
+		"target_contract", w.Job.TargetContractAddress,
+		"target_function", w.Job.TargetFunction,
+	)
+
+	// Simulate processing time
+	time.Sleep(100 * time.Millisecond)
+
+	// Simulate 95% success rate
+	return time.Now().UnixNano()%100 < 95
+}
